one more int => bigint change (#25651)
BK, run evaluationsquery

> Insert changelog entry or delete this section.

## Summary & Motivation

## How I Tested These Changes

## Changelog

> Insert changelog entry or delete this section.

+++ python_modules/dagster-graphql/dagster_graphql/schema/asset_graph.py
@@ -308,7 +308,7 @@ class GrapheneAssetNode(graphene.ObjectType):
         limit=graphene.Argument(graphene.Int),
         pipeline=graphene.Argument(GraphenePipelineSelector),
     )
-    currentAutoMaterializeEvaluationId = graphene.Int()
+    currentAutoMaterializeEvaluationId = graphene.BigInt()
     targetingInstigators = non_null_list(GrapheneInstigator)
 
     class Meta:
